added the option to use decimals to overs

+++ src/main/java/com/example/cricketApplication/models/MatchSummary.java
@@ -2,6 +2,7 @@ package com.example.cricketApplication.models;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @lombok.Getter
@@ -18,10 +19,10 @@ public class MatchSummary {
 
     private int runs;
     private int wickets;
-    private int overs;
+    private BigDecimal overs;
     private int oppositionRuns;
     private int oppositionWickets;
-    private int oppositionOvers;
+    private BigDecimal  oppositionOvers;
     private String result;
 
     @ManyToOne(fetch = FetchType.LAZY)
